feat: UserService 인터페이스 추출 및 Impl과 Transaction 로직 클래스 분리
