+++ src/main/java/springbook/service/UserService.java
@@ -1,72 +1,10 @@
 package springbook.service;
 
-import org.springframework.mail.MailSender;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.TransactionStatus;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
-import springbook.dao.UserDao;
-import springbook.domain.user.Level;
 import springbook.domain.user.User;
 
-@Service
-public class UserService {
+public interface UserService {
 
-    private UserDao userDao;
-    private UserLevelUpgradePolicy userLevelUpgradePolicy;
-    private PlatformTransactionManager transactionManager;
-    private MailSender mailSender;
+    void upgradeLevels();
 
-    public void setUserDao(UserDao userDao) {
-        this.userDao = userDao;
-    }
-
-    public void setUserLevelUpgradePolicy(UserLevelUpgradePolicy userLevelUpgradePolicy) {
-        this.userLevelUpgradePolicy = userLevelUpgradePolicy;
-    }
-
-    public void setTransactionManager(PlatformTransactionManager transactionManager) {
-        this.transactionManager = transactionManager;
-    }
-
-    public void setMailSender(MailSender mailSender) {
-        this.mailSender = mailSender;
-    }
-
-    public void upgradeLevels() {
-        TransactionStatus transaction = transactionManager.getTransaction(new DefaultTransactionDefinition());
-        try {
-            userDao.getAll()
-                    .forEach(this::upgradeLevel);
-            transactionManager.commit(transaction);
-        } catch (Exception exception) {
-            transactionManager.rollback(transaction);
-            throw exception;
-        }
-    }
-
-    protected void upgradeLevel(User user) {
-        if (userLevelUpgradePolicy.canUpgradeLevel(user)) {
-            userLevelUpgradePolicy.upgradeLevel(user);
-            userDao.update(user);
-            sendEmail(user);
-        }
-    }
-
-    private void sendEmail(User user) {
-        SimpleMailMessage simpleMailMessage = new SimpleMailMessage();
-        simpleMailMessage.setTo(user.getEmail());
-        simpleMailMessage.setFrom("useradmin@ksug.org");
-        simpleMailMessage.setSubject("Upgrade 안내");
-        simpleMailMessage.setText("사용자님의 등급이 " + user.getLevel().name());
-        mailSender.send(simpleMailMessage);
-    }
-
-    public void add(User user) {
-        if (user.getLevel() == null) {
-            user.setLevel(Level.BASIC);
-        }
-        userDao.addUser(user);
-    }
+    void add(User user);
 }
