+++ src/main/java/springbook/service/UserServiceTx.java
@@ -1,0 +1,37 @@
+package springbook.service;
+
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.TransactionStatus;
+import org.springframework.transaction.support.DefaultTransactionDefinition;
+import springbook.domain.user.User;
+
+public class UserServiceTx implements UserService {
+
+    private UserService userService;
+    private PlatformTransactionManager transactionManager;
+
+    public void setUserService(UserService userService) {
+        this.userService = userService;
+    }
+
+    public void setTransactionManager(PlatformTransactionManager transactionManager) {
+        this.transactionManager = transactionManager;
+    }
+
+    @Override
+    public void upgradeLevels() {
+        TransactionStatus transaction = transactionManager.getTransaction(new DefaultTransactionDefinition());
+        try {
+            userService.upgradeLevels();
+            transactionManager.commit(transaction);
+        } catch (RuntimeException e) {
+            transactionManager.rollback(transaction);
+            throw e;
+        }
+    }
+
+    @Override
+    public void add(User user) {
+        userService.add(user);
+    }
+}
