+++ src/main/resources/applicationContext.xml
@@ -6,12 +6,16 @@
     <bean id="userDaoJdbc" class="springbook.dao.UserDaoJdbc">
         <property name="dataSource" ref="dataSource"/>
     </bean>
-    <bean id="userService" class="springbook.service.UserService">
+    <bean id="userService" class="springbook.service.UserServiceTx">
+        <property name="userService" ref="userServiceImpl"/>
+        <property name="transactionManager" ref="transactionManager"/>
+    </bean>
+    <bean id="userServiceImpl" class="springbook.service.UserServiceImpl">
         <property name="userDao" ref="userDaoJdbc"/>
         <property name="userLevelUpgradePolicy" ref="userLevelUpgradePolicy"/>
-        <property name="transactionManager" ref="transactionManager"/>
         <property name="mailSender" ref="mailSender"/>
     </bean>
+
     <bean id="mailSender" class="org.springframework.mail.javamail.JavaMailSenderImpl">
         <property name="host" value="mail.server.com"/>
     </bean>
